consolidate functional tests
