--- liferay-workspace/tests/blade.language.test/src/testFunctional/java/com/liferay/blade/samples/language/web/test/BladeLanguageWebTest.java
@@ -1,106 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.blade.samples.language.web.test;
-
-import com.liferay.arquillian.portal.annotation.PortalURL;
-import com.liferay.portal.kernel.exception.PortalException;
-
-import java.io.File;
-
-import java.net.URL;
-
-import org.jboss.arquillian.container.test.api.Deployment;
-import org.jboss.arquillian.container.test.api.RunAsClient;
-import org.jboss.arquillian.drone.api.annotation.Drone;
-import org.jboss.arquillian.junit.Arquillian;
-import org.jboss.shrinkwrap.api.ShrinkWrap;
-import org.jboss.shrinkwrap.api.spec.JavaArchive;
-
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.FindBy;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.openqa.selenium.support.ui.WebDriverWait;
-
-/**
- * @author Liferay
- */
-@RunAsClient
-@RunWith(Arquillian.class)
-public class BladeLanguageWebTest {
-
-	@Deployment
-	public static JavaArchive create() throws Exception {
-		final File jarFile = new File(System.getProperty("jarFile"));
-
-		return ShrinkWrap.createFromZipFile(JavaArchive.class, jarFile);
-	}
-
-	@Test
-	public void testBladeSamplesLanguage() throws PortalException {
-		_webDriver.get(_portletURL.toExternalForm());
-
-		Assert.assertTrue(
-			"Portlet was not deployed", isVisible(_bladeSampleLanguagePortlet));
-		Assert.assertTrue(
-			_languageKeyFirst.getText(),
-			_languageKeyFirst.getText().contentEquals(
-				"Hello from BLADE Language Web!"));
-		Assert.assertTrue(
-			_languageKeySecond.getText(),
-			_languageKeySecond.getText().contentEquals(
-				"Hello from the BLADE Language Module!"));
-		Assert.assertTrue(
-			_languageKeyThird.getText(),
-			_languageKeyThird.getText().contentEquals(
-				"I have overridden the key from BLADE Language Module!"));
-	}
-
-	protected boolean isVisible(WebElement webelement) {
-		WebDriverWait webDriverWait = new WebDriverWait(_webDriver, 5);
-
-		try {
-			webDriverWait.until(ExpectedConditions.visibilityOf(webelement));
-
-			return true;
-		}
-		catch (org.openqa.selenium.TimeoutException te) {
-			return false;
-		}
-	}
-
-	@FindBy(xpath = "//div[contains(@id,'_com_liferay_blade_samples_language_web')]")
-	private WebElement _bladeSampleLanguagePortlet;
-
-	@FindBy(xpath = "//div[@class='portlet-body']/p[1]")
-	private WebElement _languageKeyFirst;
-
-	@FindBy(xpath = "//div[@class='portlet-body']/p[2]")
-	private WebElement _languageKeySecond;
-
-	@FindBy(xpath = "//div[@class='portlet-body']/p[3]")
-	private WebElement _languageKeyThird;
-
-	@PortalURL("com_liferay_blade_samples_language_web")
-	private URL _portletURL;
-
-	@Drone
-	private WebDriver _webDriver;
-
-}
