--- liferay-workspace/tests/blade.portlet.freemarker.test/src/testFunctional/resources/arquillian.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<arquillian xmlns="http://jboss.org/schema/arquillian"
-			xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-			xsi:schemaLocation="http://jboss.org/schema/arquillian http://jboss.org/schema/arquillian/arquillian_1_0.xsd">
-
-	<extension qualifier="webdriver">
-		<property name="browser">phantomjs</property>
-	</extension>
-
-	<extension qualifier="graphene">
-		<property name="url">http://localhost:8080</property>
-	</extension>
-</arquillian>
