--- liferay-workspace/tests/blade.portlet.osgiapi.test/bnd.bnd
@@ -1,2 +1,0 @@
-Bundle-Version: 1.0.0
--sources: true
