--- liferay-workspace/tests/blade.portlet.rendercommand.test/bnd.bnd
@@ -1,2 +1,0 @@
-Bundle-Version: 1.0.0
--sources: true
