+++ liferay-workspace/tests/blade.samples.test/bnd.bnd
@@ -1,0 +1,2 @@
+Bundle-Version: 1.0.0
+-sources: true
