+++ liferay-workspace/tests/blade.samples.test/src/testFunctional/resources/arquillian.xml
